Fix after review

+++ app/controllers/questions_controller.rb
@@ -22,8 +22,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path, notice: 'Question successfully deleted.'
+    if current_user&.author_off?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question successfully deleted.'
+    else
+      redirect_to questions_path, notice: 'You have no rigths to delete this question.'
+    end
   end
 
   private
